fixed test_settlement
the test relied on a strict order of the event, which may change
depending the timming of the test
