+++ raiden/tests/integration/test_e2e.py
@@ -9,6 +9,7 @@ from raiden.tests.utils.transfer import (
     channel,
     get_sent_transfer,
 )
+from raiden.tests.utils.events import must_contain_entry
 from raiden.tests.utils.log import get_all_state_changes, get_all_state_events
 from raiden.tests.utils.blockchain import wait_until_block
 from raiden.transfer.state_change import (
@@ -57,45 +58,6 @@ def assert_path_mediated_transfer(*transfers):
 
         assert first.recipient == second.sender, 'transfers are out-of-order'
         assert first.lock.expiration > second.lock.expiration, 'lock expiration is not decreasing'
-
-
-def check_nested_attrs(item, data):
-    for name, value in data.iteritems():
-        item_value = getattr(item, name)
-
-        if isinstance(value, dict):
-            if not check_nested_attrs(item_value, value):
-                return False
-
-        elif item_value != value:
-            return False
-
-    return True
-
-
-def must_contain_entry(item_list, type_, data):
-    """ A node might have duplicated state changes or code changes may change
-    order / quantity of the events.
-
-    The number of state changes is non-deterministic since it depends on the
-    number of retries from the protocol layer.
-
-    This is completely non-deterministic since the protocol retries depend on
-    timeouts and the cooperative scheduling of the running greenlets.
-    Additionally the order / quantity of greenlet switches will change as the
-    code evolves.
-
-    This utility checks the list of state changes for an entry of the correct
-    type with the expected data, ignoring *new* fields, repeated entries, and
-    unexpected entries.
-    """
-    # item_list may be composed of state changes or events
-    for item in item_list:
-        if isinstance(item, type_):
-            if check_nested_attrs(item, data):
-                return True
-
-    return False
 
 
 @pytest.mark.parametrize('channels_per_node', [1])
